Rename define action to let
This PR renames the "define" action to "let", which makes it clearer
the difference between the define command and this action (closes #67).

Even though this action now doesn't conflict in name with the "define"
command, it is still not allowed at the top level of the text file,
because, AFAIK, it would be a no-op there.