Expose EgglogSpan publicly and add debug to spans
This PR makes public the fields on the egglog span so that we can access
them from the Python bindings. In Python, I have been generally converting
to/from egglog sructs by creating a corresponding Python class and copying
all of their fields. To do so , it has to be public.

I also use the debug formatting to allow pretty printing in Python as
the rust struct.

See https://github.com/egraphs-good/egglog-python/pull/258 for the
corresponding PR on the Python side.