Don't rely on FixedBitset for computing length
This led to larger vectors getting added for later bindigns due to cached Masks
having a larger capacity than the initial size. This is not a problem for
data-structures like vectors but is a bit of a gotcha for bitsets.

This change also moves the `do_parallel` check from test to debug_assertions,
increasing coverage for tests in different crates.