Better change tracking during merges.
This change modifies ExecutionState to keep track of whether or not it enqueued
any changes to other tables, and then uses this tracking during the `merge_all`
method to determine whether or not to perform an extra round of the fixed-point
loop.

This results in a passing `merge-saturates` test. That test creates a case
where a round of `merge_all` causes no change to the table `bar` but has
`bar`'s merge function queue up a change to the table `baz` that simply hasn't
been merged yet.