Add walltime benchmarks
This PR will changes benchmarks to be measured with walltime (https://codspeed.io/docs/instruments/walltime) as well as instrumented.

The walltime should capture the effect of system calls, like with accessing files or lock contention with multithreading. However, it won't be as repeatable, so is run on the custom codspeed runner which is set up for repeatable benchmark timing.