Make name and partially applied fields public in function container
This PR makes a few fields public in the function container. This matches
other containers, like Vec. This is necessary to use `value_to_base` on
function values, to get the actual function name and the arguments.

I am exposing `value_to_base` to the Python API with all primitive sorts
to support that functionality while defining custom cost models, to be able
to interpret the primitive values.