Don't Panic on None Primitive in Action
Closes #697 by returning an error when running a ruleset that contains
a primitive which returns `None` instead of panicking.

This seems like generally good practice and also will let me handle Python
errors better, by re-raising them if I get an error instead of just panicking.

I was also thinking it might make sense to rename "Panic" to like "Error". This isn't a panic,
it's just when a primitive returns none, which is an expected case.