Fix Incorrect Unstable Function Behavior
Closes #738 by fixing incorrect behavior of the unstable higher order functions.
Previously, it assumed that the sorts of the partially applied args were the same as
the sort of the remaining args, or the sorts that the function sort was parametrized over.
Of course, they may be different. This fixes the bug by storing the arcsorts on the function container value.

However, the extractor currently needs to ask each sort for all of it's possible children sorts when deciding which
sorts to extract from. So we also have to store a list of all the arcsorts used in any of the partial functions on the sort.
Alternatively, we could change the extractor to not require to know from each sort all possible child sorts, but for now I have
opted to keep the interface the same and only make changes to the function sort.